backend/pyproject.toml
[project]
name = "bring-the-flow-backend"
version = "0.1.0"
description = "Backend for bring-the-flow: Hobbs-inspired flow-field art planner"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "cairosvg>=2.7",
    "pillow>=10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
